+++ lib/cerbos/output/check_resources.rb
@@ -233,7 +233,14 @@ module Cerbos
     end
 
     # User-defined output from a policy rule evaluation.
-    CheckResources::Result::Output = Output.new_class(:source, :value) do
+    CheckResources::Result::Output = Output.new_class(:action, :source, :value) do
+      # @!attribute [r] action
+      #   The action that was being evaluated when the output was produced.
+      #
+      #   Requires the Cerbos policy decision point server to be at least v0.51.
+      #
+      #   @return [String]
+
       # @!attribute [r] source
       #   The identifier of the policy rule that produced the output.
       #
@@ -246,6 +253,7 @@ module Cerbos
 
       def self.from_protobuf(output_entry)
         new(
+          action: output_entry.action,
           source: output_entry.src,
           value: output_entry.val&.to_ruby(true)
         )