+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
   Requires a policy decision point server running Cerbos 0.51+.
 
+- `Cerbos::Output::CheckResources::Result::Output#action` attribute returning the action that was being evaluated when the output was produced ([#292](https://github.com/cerbos/cerbos-sdk-ruby/pull/292))
+
+  Requires a policy decision point server running Cerbos 0.51+.
+
 ## [0.12.0] - 2025-08-12
 
 ### Added